@@ -19,8 +19,6 @@
     "@dfinity/agent": "0.10.3",
     "@dfinity/candid": "0.10.3",
     "@dfinity/principal": "0.10.3",
-    "@material-ui/core": "4.6.1",
-    "@material-ui/icons": "4.5.1",
     "assert": "2.0.0",
     "buffer": "6.0.3",
     "copy-webpack-plugin": "^10.2.4",
@@ -45,10 +43,13 @@
     "last 2 edge version"
   ],
   "dependencies": {
-    "@material-ui/core": "4.6.1",
-    "@material-ui/icons": "4.5.1",
-    "react": "^16.8.2",
-    "react-dom": "^16.8.2",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.5",
+    "react": "17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.0.2",
     "react-scripts": "5.0.0",
     "uuid": "8.3.2"
   }
