@@ -1,5 +1,5 @@
-import React from "react";
-import DeleteIcon from "@material-ui/icons/Delete";
+import React from 'react';
+import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 
 function Note(props) {
   function handleClick() {
@@ -11,7 +11,7 @@ function Note(props) {
       <h1>{props.title}</h1>
       <p>{props.content}</p>
       <button onClick={handleClick}>
-        <DeleteIcon />
+        <DeleteForeverIcon />
       </button>
     </div>
   );
